requirements.txt
-i https://pypi.org/simple
aiosqlite==0.19.0 ; python_version >= '3.7'
alembic==1.10.4 ; python_version >= '3.7'
anyio==3.6.2 ; python_full_version >= '3.6.2'
apprise==1.3.0 ; python_version >= '3.6'
asgi-lifespan==2.1.0 ; python_version >= '3.7'
asyncpg==0.27.0 ; python_full_version >= '3.7.0'
attrs==23.1.0 ; python_version >= '3.7'
autograd==1.5
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
cloudpickle==2.2.1 ; python_version >= '3.6'
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coolname==2.2.0
croniter==1.3.14 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==40.0.2 ; python_version >= '3.6'
dateparser==1.1.8 ; python_version >= '3.7'
db-dtypes==1.1.1
dill==0.3.6 ; python_version >= '3.7'
docker==6.1.2 ; python_version >= '3.7'
fastapi==0.95.1 ; python_version >= '3.7'
fsspec==2023.5.0 ; python_version >= '3.8'
future==0.18.3 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc]==2.11.0 ; python_version >= '3.7'
google-api-python-client==2.86.0 ; python_version >= '3.7'
google-auth==2.18.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-cloud-bigquery[pandas]==3.10.0
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.9.0 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.5.0 ; python_version >= '3.7'
googleapis-common-protos==1.59.0 ; python_version >= '3.7'
greenlet==2.0.2
griffe==0.27.5 ; python_version >= '3.7'
grpc-google-iam-v1==0.12.6 ; python_version >= '3.7'
grpcio==1.54.2 ; python_version >= '3.7'
grpcio-status==1.54.2
h11==0.14.0 ; python_version >= '3.7'
h2==4.1.0
hpack==4.0.0 ; python_full_version >= '3.6.1'
httpcore==0.17.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx[http2]==0.24.0 ; python_version >= '3.7'
hyperframe==6.0.1 ; python_full_version >= '3.6.1'
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
jsonpatch==1.32 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jsonpointer==2.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==4.17.3 ; python_version >= '3.7'
kubernetes==26.1.0 ; python_version >= '3.6'
lifetimes==0.11.3
mako==1.2.4 ; python_version >= '3.7'
markdown==3.4.3 ; python_version >= '3.7'
markdown-it-py==2.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
msal==1.22.0
numpy==1.24.3
oauthlib==3.2.2 ; python_version >= '3.6'
orjson==3.8.12 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.1
pathspec==0.11.1 ; python_version >= '3.7'
patsy==0.5.3
pendulum==2.1.2
prefect==2.10.9
prefect-airbyte==0.2.0
prefect-gcp[secret_manager]==0.4.1
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.23.0 ; python_version >= '3.7'
pyarrow==12.0.0
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==1.10.7 ; python_version >= '3.7'
pygments==2.15.1 ; python_version >= '3.7'
pyjwt[crypto]==2.7.0 ; python_version >= '3.7'
pyodbc==4.0.39
pyopenssl==23.1.1
pyparsing==3.0.9 ; python_version >= '3.1'
pyrsistent==0.19.3 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==8.0.1 ; python_version >= '3.7'
pytz==2023.3
pytzdata==2020.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0 ; python_version >= '3.6'
readchar==4.0.5 ; python_version >= '3.7'
regex==2023.5.5 ; python_version >= '3.6'
requests==2.30.0
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.3.5 ; python_full_version >= '3.7.0'
rsa==4.9 ; python_version >= '3.6'
scikit-learn==1.2.2
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy[asyncio]==1.4.48
starlette==0.26.1 ; python_version >= '3.7'
statsmodels==0.14.0
text-unidecode==1.3
threadpoolctl==3.1.0 ; python_version >= '3.6'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typer==0.9.0 ; python_version >= '3.6'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
tzlocal==5.0 ; python_version >= '3.7'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.22.0 ; python_version >= '3.7'
websocket-client==1.5.1 ; python_version >= '3.7'
websockets==11.0.3 ; python_version >= '3.7'
xgboost==1.7.5
